refactor: centered on grammarReview.py
